Modularize tests (so unrelated tests break less frequently)
