+++ test/rsvd.jl
@@ -1,6 +1,8 @@
 using IterativeSolvers
 using FactCheck
 
+srand(1234321)
+
 for rsvd in [IterativeSolvers.rsvdfact]
     facts(rsvd) do
 
@@ -51,7 +53,7 @@ for rsvd in [IterativeSolvers.rsvdfact]
             A = [1. 2 3; 4 5 6; 7 8 9]
             @fact size(IterativeSolvers.rrange_adaptive(A, 3, 1e-3)) --> (3,2)
         end
-        
+
         context("rrange") do
             A = [1. 2 3; 4 5 6; 7 8 9]
             @fact_throws ArgumentError IterativeSolvers.rrange(A, 20)
